API/requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==2.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
Pillow==7.0.0
psycopg2==2.8.5
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2019.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
touch==2019.4.13
Twisted==19.10.0
txaio==20.1.1
values==2019.4.13
whitenoise==5.0.1
zope.interface==4.7.1